Implement basic free fall detection
The gravity sensor appears to always re-introduce the force of gravity
so as to provide the gravity vector in it's true (absolute) meaning, that
is, not relative to the device's acceleration.

The regular acceleromter, however, seems to provide us with what we want.
Surprisingly, however, the error threshold required to detect a fall
seems quite high (1.0f m/s^2) and there seem to also be calibration issues
(or gravity is just stronger here in Berlin?).